+++ tests/Fixtures/Resolver/Primary/AmbiguousSpider.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2024 Kai Sassnowski
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ *
+ * @see https://github.com/roach-php/laravel
+ */
+
+namespace RoachPHP\Laravel\Tests\Fixtures\Resolver\Primary;
+
+/**
+ * @internal
+ */
+final class AmbiguousSpider
+{
+}